backend/requirements.txt
# Web framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0

# ORM + Data models
sqlmodel>=0.0.22,<1.0

# Settings management (Pydantic v2)
pydantic-settings>=2.0,<3.0

# Encryption
cryptography>=43,<44
pyrage>=1.1,<2.0
argon2-cffi>=23,<24

# Shamir's Secret Sharing (SLIP-39)
shamir-mnemonic>=0.3,<1.0

# Vector database client
qdrant-client>=1.12,<2.0

# HTTP client (for Ollama API calls)
httpx>=0.27,<1.0

# HTML content extraction (URL ingestion)
readability-lxml>=0.8,<1.0
lxml[html_clean]>=5.0,<6.0

# Git operations
GitPython>=3.1,<4.0

# Image processing
Pillow>=11,<12

# File type detection
python-magic>=0.4.27,<1.0

# Auth (JWT)
python-jose[cryptography]>=3.3,<4.0

# Multipart file uploads
python-multipart>=0.0.9,<1.0

# Testing
pytest>=8,<9
pytest-asyncio>=0.24,<1.0
